requirements.txt
snowflake-connector-python
openpyxl
boto3
anthropic
